Finish moc test 3

+++ Exercises/MockTestsExercises/Test2/simulaprova2.1/src/main/java/Try4/models/Gasto.java
@@ -5,7 +5,7 @@ import Try4.persistence.Entidade;
 import java.util.Objects;
 
 public sealed class Gasto implements Entidade<Long> permits GastoSaude, GastoEducacao{
-    public static double idCount = 1;
+    public static long idCount = 1;
     private final long id;
     private final String descricao;
     private final double valor;
